pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lanhu-cli"
version = "1.0.0"
description = "Standalone CLI for Lanhu design platform — scrape pages, designs, messages without MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "httpx[socks]>=0.27.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
]

[project.scripts]
lanhu = "lanhu_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
